backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
yfinance==1.2.1
ta==0.11.0
httpx==0.28.1
python-dotenv==1.2.2
pytest==8.3.5
pytest-asyncio==1.3.0
pandas==3.0.2
